requirements.txt
Flask==1.1.1
request==2019.4.13
jsonify==0.5
dnspython==1.16.0
pymongo==3.10.1
bcrypt==3.1.7
